--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -11,7 +11,6 @@ jobs:
     strategy:
       matrix:
         python:
-          - "3.7"
           - "3.8"
           - "3.9"
           - "3.10"
@@ -46,7 +45,6 @@ jobs:
           - ubuntu-latest
           - macos-latest
         python:
-          - "3.7"
           - "3.8"
           - "3.9"
           - "3.10"
@@ -78,7 +76,7 @@ jobs:
       - uses: actions/checkout@v3
       - uses: actions/setup-python@v4
         with:
-          python-version: '3.7'
+          python-version: '3.8'
       - name: Install cleanlab
         run: |
           python -m  pip install --upgrade pip