--- a/.github/workflows/ci.yml
+++ b/.github/workflows/ci.yml
@@ -14,6 +14,7 @@ jobs:
           - "3.8"
           - "3.9"
           - "3.10"
+          - "3.11"
     steps:
       - uses: actions/checkout@v3
       - uses: actions/setup-python@v4
@@ -26,7 +27,7 @@ jobs:
       - name: Install cleanlab
         run: pip install -e ".[all]"
       - name: Install development dependencies
-        run: pip install -r requirements-dev.txt
+        run: pip install -vr requirements-dev.txt
       - name: Overwrite tensorflow version on Windows
         run: |
           pip uninstall -y tensorflow
@@ -48,6 +49,7 @@ jobs:
           - "3.8"
           - "3.9"
           - "3.10"
+          - "3.11"
     steps:
       - uses: actions/checkout@v3
       - uses: actions/setup-python@v4
@@ -60,7 +62,7 @@ jobs:
       - name: Install cleanlab
         run: pip install -e ".[all]"
       - name: Install development dependencies
-        run: pip install -r requirements-dev.txt
+        run: pip install -vr requirements-dev.txt
       - name: Install fasttext for non-Windows machines
         run: |
           pip install fasttext